+++ stdlib/tests/collections/mod.rs
@@ -1,1 +1,2 @@
 mod mmr;
+mod smt;